split up parts into own files
